+++ back/src/main/java/com/example/gmall/service/impl/SellerDashboardServiceImpl.java
@@ -56,7 +56,7 @@ public class SellerDashboardServiceImpl implements SellerDashboardService {
                         .receiverName(o.getReceiverName())
                         .totalPrice(o.getTotalPrice())
                         .status(o.getStatus())
-                        .statusLabel(switch (o.getStatus()) {
+                        .statusLabel(switch (o.getStatus().intValue()) {
                             case 0 -> "결제전";
                             case 1 -> "상품준비중";
                             case 2 -> "배송중";
